:tada: feat: add reauthorize session modal (ref #928)
